Remove time.Sleep and refactor fixture function to enable hook
