+++ api/app_test.go
@@ -63,6 +63,7 @@ var _ = Describe("API Application", func() {
 		var err error
 		testDb, err = GetTestDB()
 		Expect(err).NotTo(HaveOccurred())
+		fixtures.ConfigureAndStartGoWorkers()
 	})
 
 	Describe("App Struct", func() {
@@ -81,7 +82,6 @@ var _ = Describe("API Application", func() {
 			Expect(err).NotTo(HaveOccurred())
 
 			app := GetDefaultTestApp()
-			app.NonblockingStartWorkers()
 
 			appGame, err := app.GetGame(nil, game.PublicID)
 			Expect(err).NotTo(HaveOccurred())
@@ -94,7 +94,6 @@ var _ = Describe("API Application", func() {
 			Expect(err).NotTo(HaveOccurred())
 
 			app := GetDefaultTestApp()
-			app.NonblockingStartWorkers()
 
 			appGame, err := app.GetGame(nil, game.PublicID)
 			Expect(err).NotTo(HaveOccurred())
@@ -110,7 +109,6 @@ var _ = Describe("API Application", func() {
 			Expect(err).NotTo(HaveOccurred())
 
 			app := GetDefaultTestApp()
-			app.NonblockingStartWorkers()
 
 			hooks := app.GetHooks(context.Background())
 			Expect(len(hooks[gameID])).To(Equal(2))
@@ -129,7 +127,6 @@ var _ = Describe("API Application", func() {
 			responses := startRouteHandler([]string{"/created", "/created2"}, 52525)
 
 			app := GetDefaultTestApp()
-			app.NonblockingStartWorkers()
 
 			resultingPayload := map[string]interface{}{
 				"success":  true,
@@ -156,7 +153,6 @@ var _ = Describe("API Application", func() {
 			)
 
 			app := GetDefaultTestApp()
-			app.NonblockingStartWorkers()
 
 			resultingPayload := map[string]interface{}{
 				"url":      "http://some-url.com",
@@ -191,7 +187,6 @@ var _ = Describe("API Application", func() {
 			responses := startRouteHandler([]string{fmt.Sprintf("/created/%s", hooks[0].GameID)}, 52525)
 
 			app := GetDefaultTestApp()
-			app.NonblockingStartWorkers()
 
 			resultingPayload := map[string]interface{}{
 				"success":  true,
@@ -215,7 +210,6 @@ var _ = Describe("API Application", func() {
 			responses := startRouteHandler([]string{fmt.Sprintf("/1/créated/%s", hooks[0].GameID)}, 52525)
 
 			app := GetDefaultTestApp()
-			app.NonblockingStartWorkers()
 
 			resultingPayload := map[string]interface{}{
 				"success":        true,
@@ -242,7 +236,6 @@ var _ = Describe("API Application", func() {
 			responses := startRouteHandler([]string{fmt.Sprintf("/invalid-webhook-request/%s", hooks[0].GameID)}, 52525)
 
 			app := GetDefaultTestApp()
-			app.NonblockingStartWorkers()
 
 			resultingPayload := map[string]interface{}{
 				"success": true,
