+++ models/mongo_worker.go
@@ -2,6 +2,8 @@ package models
 
 import (
 	"context"
+	"encoding/json"
+	"errors"
 	"fmt"
 
 	"github.com/jrallison/go-workers"
@@ -50,6 +52,22 @@ func (w *MongoWorker) PerformUpdateMongo(m *workers.Msg) {
 	clanID := data["clanID"].(string)
 
 	w.updateClanIntoMongoDB(ctx, game, op, clan, clanID)
+}
+
+// InsertGame creates a game inside Mongo
+func (w *MongoWorker) InsertGame(ctx context.Context, gameID string, clan *Clan) error {
+	clanWithNamePrefixes := clan.NewClanWithNamePrefixes()
+	clanJSON, err := json.Marshal(clanWithNamePrefixes)
+	if err != nil {
+		return errors.New("Could not serialize clan")
+	}
+
+	var clanMap map[string]interface{}
+	json.Unmarshal(clanJSON, &clanMap)
+
+	w.updateClanIntoMongoDB(ctx, gameID, "update", clanMap, clan.PublicID)
+
+	return nil
 }
 
 func (w *MongoWorker) updateClanIntoMongoDB(
